# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{authUserId}/tier-history:
    get:
      operationId: adminGetUserTierHistory
      tags: [admin]
      summary: Subscription tier history for any user (support lookup)
      description: |
        Returns every tier transition for the specified user in reverse-chronological order.
        Intended for support staff to answer billing questions without requiring the customer
        to be present or granting direct database access.
        Requires the `X-Admin-Key` header to match the server's `ADMIN_API_KEY` environment variable.
      parameters:
        - name: authUserId
          in: path
          required: true
          schema:
            type: string
          description: The Auth user ID of the user to look up.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of records to return (1–100, default 50).
      security:
        - AdminKey: []
      responses:
        "200":
          description: Tier transition history for the specified user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TierHistoryResponse"
        "401":
          description: Missing X-Admin-Key header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Admin API not configured on this server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /legal/document:
    get:
      operationId: legalGetDocument
      tags: [legal]
      summary: Full text of a legal document in a given locale
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [privacy_policy, terms_of_service]
        - name: locale
          in: query
          required: false
          schema:
            type: string
            enum: [en, az, ru, ar, zh, es]
      responses:
        "200":
          description: Document text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalDocumentResponse"
        "400":
          description: Invalid type or locale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /legal/me:
    get:
      operationId: legalMyAcceptances
      tags: [legal]
      summary: Latest acceptance per document for the signed-in user
      responses:
        "200":
          description: Acceptance state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalAcceptancesState"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /legal/accept:
    post:
      operationId: legalAccept
      tags: [legal]
      summary: Record acceptance of one or more legal documents at their current versions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalAcceptInput"
      responses:
        "200":
          description: Updated acceptance state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalAcceptancesState"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Secret key required for all /admin/* endpoints. Set via the ADMIN_API_KEY environment variable on the server.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TierTransitionEntry:
      type: object
      description: A single tier change event in the user's subscription history.
      properties:
        id:
          type: integer
        fromTier:
          type: string
        toTier:
          type: string
        triggeredBy:
          type: string
          description: '"webhook" or "sync-tier"'
        eventType:
          oneOf:
            - type: string
            - type: "null"
          description: Raw RevenueCat event type (e.g. "INITIAL_PURCHASE"). Null for sync-tier entries.
        createdAt:
          type: string
          format: date-time
      required: [id, fromTier, toTier, triggeredBy, eventType, createdAt]
    TierHistoryResponse:
      type: object
      properties:
        transitions:
          type: array
          items:
            $ref: "#/components/schemas/TierTransitionEntry"
        total:
          type: integer
          description: Total number of transitions returned in this response.
      required: [transitions, total]
...
    LegalDocumentType:
      type: string
      enum: [privacy_policy, terms_of_service]
    LegalLocale:
      type: string
      enum: [en, az, ru, ar, zh, es]
...
    LegalDocumentResponse:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/LegalDocumentType"
        version:
          type: string
        locale:
          $ref: "#/components/schemas/LegalLocale"
        title:
          type: string
        body:
          type: string
        authoritativeNotice:
          type:
            - string
            - "null"
          description: Convenience-translation notice when the requested locale is not the authoritative one (English).
      required: [type, version, locale, title, body, authoritativeNotice]
    LegalAcceptanceRecord:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/LegalDocumentType"
        acceptedVersion:
          type:
            - string
            - "null"
        currentVersion:
          type: string
        acceptedAt:
          type:
            - string
            - "null"
        locale:
          type:
            - string
            - "null"
        upToDate:
          type: boolean
      required: [type, acceptedVersion, currentVersion, acceptedAt, locale, upToDate]
    LegalAcceptancesState:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/LegalAcceptanceRecord"
        allUpToDate:
          type: boolean
      required: [documents, allUpToDate]
    LegalAcceptItem:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/LegalDocumentType"
        version:
          type: string
      required: [type, version]
    LegalAcceptInput:
      type: object
      properties:
        locale:
          $ref: "#/components/schemas/LegalLocale"
        documents:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/LegalAcceptItem"
      required: [locale, documents]